Update README.md
In this readme, I have provided step-by-step instructions on how to use the "create_vultr_vps.sh" script for automating the process of creating a new virtual private server (VPS) instance on the Vultr platform using the Vultr API. The instructions include how to customize the script, how to make the script executable, how to run the script, how to check that the VPS has been created, and how to use the jq command. You also described how to select the desired operating system, plan, and region for your VPS instance using Vultr's API. Additionally, you provided installation instructions for jq and shared that the script is open-source and can be shared on Github.